Moved README.md to docs/developer-notes/developer-scratchpad.md
When I started on my README.md I intended it to be a detailed set of notes to explain the problem, my approach, and my solution. It is presently a disorganized jumble. It is also a little long (not real long). as a consequence if I organize it and add the content directed by the programming assignment instructions it will be too long.

So for now I am moving it to docs/developer-notes/developer-scratchpad.md . I will eventually write a new README.md and some other documentation files and notes. I may cannibalize that docs/developer-notes/developer-scratchpad.md which is likely or I may delete it entirely, which is also likely.